requirements.txt
requests
Pillow